+++ packages/website/src/common/SiteDetails/index.tsx
@@ -19,7 +19,7 @@ import Surveys from "./Surveys";
 import CardTitle, { Value } from "./CardTitle";
 import CombinedCharts from "../Chart/CombinedCharts";
 import type { Range, Reef, SpotterData } from "../../store/Reefs/types";
-import { locationCalculator } from "../../helpers/locationCalculator";
+import { locationCalculator } from "../../helpers/map";
 import { formatNumber } from "../../helpers/numberUtils";
 import { sortByDate } from "../../helpers/sortDailyData";
 import { SurveyListItem, SurveyPoint } from "../../store/Survey/types";
