+++ packages/website/src/helpers/map.ts
@@ -1,3 +1,4 @@
+import { latLngBounds, LatLngBounds } from "leaflet";
 import type { Point, Polygon, Position } from "../store/Reefs/types";
 
 export const locationCalculator = (point: Point | Polygon): Position => {
@@ -14,4 +15,16 @@ export const locationCalculator = (point: Point | Polygon): Position => {
   const latMean = latArray.reduce((a, b) => a + b) / len;
 
   return [lngMean, latMean];
+};
+
+export const mapBounds = (polygon: Polygon): LatLngBounds => {
+  const latArr = polygon.coordinates[0].map((coord) => coord[1]);
+  const lngArr = polygon.coordinates[0].map((coord) => coord[0]);
+
+  const north = Math.max(...latArr);
+  const south = Math.min(...latArr);
+  const east = Math.max(...lngArr);
+  const west = Math.min(...lngArr);
+
+  return latLngBounds([south, west], [north, east]);
 };
